use Stimulus as suggested
