+++ app/javascript/controllers/carte_inscrits/departements-version-simplifiee.json
