+++ app/javascript/controllers/carte_inscrits/pop2021.json
