+++ app/views/pages/carte_inscrits.erb
@@ -20,35 +20,31 @@
 
 <h1><%= VaccinationCenter.count %> centres sur CovidListe, et des volontaires dans tous les d&eacute;partements</h1>
 
-<div id="map-container">
+<div
+  id="map-container"
+  data-controller="carte-inscrits--map"
+  data-map='<%=
+    vaccination_centers = VaccinationCenter.all.filter { |v| v.confirmed? }.map do |v|
+      {
+        'coordinates' => [v.lon, v.lat],
+        'nom' => v.name,
+      }
+    end
+    users_by_dept = Rails.cache.fetch("users_by_dept", expires_in: 1.hours) do
+      User.all
+        .filter { |u| u.confirmed? }
+        .group_by { |u| u.departement_code }
+        .reject { |departement_code, _| departement_code.nil? }
+        .transform_values(&:count)
+        .reject { |_, count| count < 10 }
+    end
+    objects = {
+      'vaccinationCenters' => vaccination_centers,
+      'usersByDept' => users_by_dept
+    }
+    objects.to_json
+  %>'
+>
   <div id="map"></div>
   <canvas id="deck-canvas"></canvas>
 </div>
-
-<script language="javascript">
-  window.addEventListener('load', () => {
-    const data = <%=
-      vaccination_centers = VaccinationCenter.all.filter { |v| v.confirmed? }.map do |v|
-        {
-          'coordinates' => [v.lon, v.lat],
-          'nom' => v.name,
-        }
-      end
-      users_by_dept = Rails.cache.fetch("users_by_dept", expires_in: 1.hours) do
-        User.all
-          .filter { |u| u.confirmed? }
-          .group_by { |u| u.departement_code }
-          .reject { |departement_code, _| departement_code.nil? }
-          .transform_values(&:count)
-          .reject { |_, count| count < 10 }
-      end
-      objects = {
-        'vaccinationCenters' => vaccination_centers,
-        'usersByDept' => users_by_dept
-      }
-      raw(objects.to_json)
-    %>;
-
-    const _ = createCarteInscrits(data);
-  });
-</script>
